gradient hms/correlations/average filter

+++ scripts/heatmap.py
@@ -73,7 +73,7 @@ def heatmap_generator(cfg, img_addr, attention_type, saliency, attribution_metho
     return saliency_map
             
 
-def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attention_type, SIM_PATH, MAIN_CSV_PATH, HEATMAP_FOLDER_PATH, HEATMAP_CSV_PATH, HEATMAP_IMG_PATH, MODE='new_calc'):
+def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attention_type, SIM_PATH, MAIN_CSV_PATH, HEATMAP_FOLDER_PATH, HEATMAP_CSV_PATH, HEATMAP_IMG_PATH, HEATMAP_IMG_GRADIENT_PATH, MODE='new_calc'):
     """
     Given a simulation by Udacity, the script reads the corresponding image paths from the csv and creates a heatmap for
     each driving image. The heatmap is created with the SmoothGrad algorithm available from tf-keras-vis
@@ -82,7 +82,7 @@ def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attent
 
     """
 
-    cprintf(f"Computing attention heatmaps for simulation \"{simulation_name}\" using \"{attention_type}\" of run id \"{run_id}\"", 'l_cyan')
+    cprintf(f"Computing attention heatmaps in {MODE} mode for simulation \"{simulation_name}\" using \"{attention_type}\" of run id \"{run_id}\"", 'l_cyan')
 
     # load the image file paths from main csv
     main_data = pd.read_csv(MAIN_CSV_PATH)
@@ -108,7 +108,7 @@ def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attent
         ('Saliency', 'saliency'),
         ('Guided_BP', 'guidedbp'),
         ('SmoothGrad_2', 'smoothgrad'),
-        ('Gradient*Input', 'grad*input'),
+        ('Gradient-Input', 'grad*input'),   
         ('IntegGrad', 'intgrad'),
         ('Epsilon_LRP', 'elrp')
     ]
@@ -162,114 +162,77 @@ def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attent
         prev_hm = gradient = np.zeros((80, 160))
 
         missing_heatmaps = 0
+        missing_gradients = 0
 
         cprintf(f'Generating heatmaps and loss scores/plots...', 'l_cyan')
         for idx, img_addr in enumerate(tqdm(data)):
-            if MODE != 'new_calc':
-                # double-check if every heatmap actually exists
-                img_address = img_addr.split('/')[-1]
-                img_name = os.path.basename(os.path.normpath(img_address))
-                hm_name = "htm-" + attention_type.lower() + '-' + img_name
+            # replace drive letter if database is copied
+            # cprintf(f'{img_addr}', 'l_red')
+            if os.path.splitdrive(img_addr)[0] != os.path.splitdrive(os.getcwd())[0]:
+                if '\\ThirdEye\\ase22\\' in img_addr:
+                    img_addr = img_addr.replace(os.path.splitdrive(img_addr)[0], os.path.splitdrive(os.getcwd())[0])
+            # if 'C:\\ThirdEye\\ase22\\' in img_addr:
+            #     img_addr = img_addr.replace('C:\\ThirdEye\\ase22\\', 'D:\\ThirdEye\\ase22\\')
+            # Double-check if every heatmap/gradient actually exists
+            file_name = Path(img_addr).stem
+            hm_name = "htm-" + attention_type.lower() + '-' + file_name + '.JPG'
+            # img_address = img_addr.split('/')[-1]
+            # img_name = os.path.basename(os.path.normpath(img_address))
+            # hm_name = "htm-" + attention_type.lower() + '-' + img_name
+            
+            if MODE == 'gradient_calc' or MODE == 'csv_missing':
                 if hm_name in os.listdir(HEATMAP_IMG_PATH):
-                    # store the addresses for the heatmap csv file
-                    heatmap_path = os.path.join(HEATMAP_IMG_PATH, hm_name)
-                    list_of_image_paths.append(heatmap_path)
-                    continue
+                    if MODE == 'csv_missing':
+                        heatmap_path = os.path.join(HEATMAP_IMG_PATH, hm_name)
+                        list_of_image_paths.append(heatmap_path)
+                        continue
                 else:
                     missing_heatmaps += 1
+            elif MODE == 'heatmap_calc':
+                if not (hm_name in os.listdir(HEATMAP_IMG_GRADIENT_PATH)):
+                    missing_gradients += 1
             
-            saliency_map = heatmap_generator(cfg, img_addr, attention_type, saliency, attribution_methods, self_driving_car_model)
-            # # convert Windows path, if needed
-            # if "\\\\" in img:
-            #     img = img.replace("\\\\", "/")
-            # elif "\\" in img:
-            #     img = img.replace("\\", "/")
+            # generate or load heatmap
+            if  MODE == 'new_calc' or MODE == 'heatmap_calc':
+                saliency_map = heatmap_generator(cfg, img_addr, attention_type, saliency, attribution_methods, self_driving_car_model)
+                if cfg.METHOD == 'thirdeye':
+                    # compute average of the heatmap
+                    average = np.average(saliency_map)
+            elif MODE == 'gradient_calc':
+                # load already generated heatmap
+                saliency_map = mpimg.imread(os.path.join(HEATMAP_IMG_PATH, hm_name))
 
-            # # load image
-            # x = mpimg.imread(img)
-
-            # # preprocess image
-            # x = utils.resize(x).astype('float32')
-
-            # # compute heatmap image
-            # saliency_map = None
-            # if attention_type == "SmoothGrad":
-            #     saliency_map = saliency(score_when_decrease, x, smooth_samples=20, smooth_noise=0.20)
-            # elif attention_type == "GradCam++":
-            #     saliency_map = saliency(score_when_decrease, x, penultimate_layer=-1)
-            # else:
-            #     attributions_orig = collections.OrderedDict()
-            #     attributions_sparse = collections.OrderedDict()
-            #     if attention_type not in attribution_methods:
-            #         raise ValueError(Fore.RED + f'Unknown heatmap computation method {attention_type} given.' + Fore.RESET)
-            #     else:
-            #         xs = x
-            #         xs = xs[np.newaxis is None,:,:,:]
-            #         xs_tensor = tensor(xs)
-
-            #         T = self_driving_car_model(xs_tensor)
-
-            #         with tf.GradientTape(persistent=True) as tape:
-            #             tape.watch(xs_tensor)
-            #             y_pred = self_driving_car_model(xs_tensor)
-
-            #         # DeepExplain initialization
-            #         with DeepExplain() as de:
-            #             for k, v in attribution_methods.items():
-
-            #                 # Explanation using DeepExplain
-            #                 attribution = de.explain(v, T, xs, xs_tensor, y_pred, tape, self_driving_car_model)
-                            
-            #                 # Preprocessing based on the method used
-            #                 if 'RectGrad' in k:
-            #                     if cfg.SPARSE_ATTRIBUTION:
-            #                         attributions_sparse[k] = preprocess(attribution, 90, 99.5)
-            #                         saliency_map = attributions_sparse[k]
-            #                     else:
-            #                         attributions_orig[k] = preprocess(attribution, 0.5, 99.5)
-            #                         saliency_map = attributions_orig[k]
-            #                 else:
-            #                     if cfg.SPARSE_ATTRIBUTION:
-            #                         attributions_sparse[k] = preprocess(attribution, 95, 99.5)
-            #                         saliency_map = attributions_sparse[k]
-            #                     else:
-            #                         attributions_orig[k] = preprocess(attribution, 0.5, 99.5)
-            #                         saliency_map = attributions_orig[k]
-
-            if cfg.METHOD == 'thirdeye':
-                # compute average of the heatmap
-                average = np.average(saliency_map)
-
-                # compute gradient of the heatmap
+            # compute gradient of the heatmap
+            if MODE == 'new_calc' or MODE == 'gradient_calc':
                 if idx == 0:
                     gradient = 0
                 else:
                     gradient = abs(prev_hm - saliency_map)
-                average_gradient = np.average(gradient)
                 prev_hm = saliency_map
 
-            # store the heatmaps
-            if not os.path.exists(HEATMAP_IMG_PATH):
-                cprintf(f'Heatmap image folder does not exist. Creating folder ...' ,'l_blue')
-                os.makedirs(HEATMAP_IMG_PATH)
-            # if cfg.SPARSE_ATTRIBUTION:
-            #     SPARSE_HEATMAP_PATH = os.path.join(HEATMAP_FOLDER_PATH, 'SPARSE_IMG')
-            #     if not os.path.exists(SPARSE_HEATMAP_PATH):
-            #         cprintf(f'Heatmap image folder does not exist. Creating folder ...' ,'l_blue')
-            #         os.makedirs(SPARSE_HEATMAP_PATH)
-            file_name = Path(img_addr).stem
-            file_name = "htm-" + attention_type.lower() + '-' + file_name + '.JPG'
-            # file_name = img_addr.split('/')[-1]
-            # file_name = "htm-" + attention_type.lower() + '-' + file_name
-            path_name = os.path.join(HEATMAP_IMG_PATH, file_name)
-            # if attention_type == "SmoothGrad" or attention_type == "GradCam++":
-            mpimg.imsave(path_name, np.squeeze(saliency_map))
-            # else:
-            #     attribution_map = np.squeeze(attributions_orig[attention_type])
-            #     v, cmap = pixel_range(attribution_map)
-            #     plt.imshow(attribution_map, vmin=v[0], vmax=v[1], cmap=cmap)
-            #     plt.savefig(path_name)
+                if cfg.METHOD == 'thirdeye':
+                    average_gradient = np.average(gradient)
 
+            # store the heatmaps
+            if  MODE == 'new_calc' or MODE == 'heatmap_calc':
+                if not os.path.exists(HEATMAP_IMG_PATH):
+                    cprintf(f'Heatmap image folder does not exist. Creating folder ...' ,'l_blue')
+                    os.makedirs(HEATMAP_IMG_PATH)
+
+                path_name = os.path.join(HEATMAP_IMG_PATH, hm_name)
+                mpimg.imsave(path_name, np.squeeze(saliency_map))
+            
+            # store the heatmap gradients
+            elif MODE == 'new_calc' or MODE == 'gradient_calc':
+                if idx != 0:
+                    if not os.path.exists(HEATMAP_IMG_GRADIENT_PATH):
+                        cprintf(f'Heatmap gradient folder does not exist. Creating folder ...' ,'l_blue')
+                        os.makedirs(HEATMAP_IMG_GRADIENT_PATH)
+                    
+                    path_name = os.path.join(HEATMAP_IMG_GRADIENT_PATH, hm_name)
+                    mpimg.imsave(path_name, np.squeeze(gradient))
+
+            path_name = os.path.join(HEATMAP_IMG_PATH, hm_name)
             list_of_image_paths.append(path_name)
 
             if cfg.METHOD == 'thirdeye':
@@ -308,7 +271,8 @@ def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attent
                 plt.savefig(AVG_GRAD_PLOT_PATH)
                 #plt.show()
         else:
-            if missing_heatmaps > 0:
+            if (missing_heatmaps > 0) and (missing_heatmaps != NUM_OF_FRAMES):
+                cprintf(f'{missing_heatmaps}', 'l_red')
                 # remove hm folder
                 shutil.rmtree(HEATMAP_FOLDER_PATH)
                 if cfg.METHOD == 'thirdeye':
@@ -317,7 +281,18 @@ def compute_heatmap(cfg, nominal, simulation_name, NUM_OF_FRAMES, run_id, attent
                     os.remove(AVG_PLOT_PATH)
                     os.remove(AVG_GRAD_SCORE_PATH)
                     os.remove(AVG_GRAD_PLOT_PATH)
-                raise ValueError("Error in number of frames of aved heatmaps. Removing all the heatmaps! Please rerun the code.")
+                raise ValueError("Error in number of frames of saved heatmaps. Removing all the heatmaps! Please rerun the code.")
+            elif (missing_gradients > 0) and (missing_gradients != NUM_OF_FRAMES-1):
+                # remove GRADIENT folder
+                shutil.rmtree(HEATMAP_IMG_GRADIENT_PATH)
+                if cfg.METHOD == 'thirdeye':
+                    # remove scores and plots
+                    os.remove(AVG_SCORE_PATH)
+                    os.remove(AVG_PLOT_PATH)
+                    os.remove(AVG_GRAD_SCORE_PATH)
+                    os.remove(AVG_GRAD_PLOT_PATH)
+                raise ValueError("Error in number of frames of saved gradients. Removing all the gradients! Please rerun the code.")
+                
 
         # save as csv in heatmap folder
         try:
